+++ folder_revision.go
@@ -2,7 +2,7 @@ package wl
 
 // FolderRevision contains information about the revision of folder.
 type FolderRevision struct {
-	ID         uint   `json:"id" yaml:"id"`
+	ID         uint64 `json:"id" yaml:"id"`
 	TypeString string `json:"type" yaml:"type"`
-	Revision   uint   `json:"revision" yaml:"revision"`
+	Revision   uint64 `json:"revision" yaml:"revision"`
 }
