+++ root.go
@@ -2,7 +2,7 @@ package wl
 
 // Root contains information about the root of the object hierarchy.
 type Root struct {
-	ID       uint `json:"id" yaml:"id"`
-	Revision uint `json:"revision" yaml:"revision"`
-	UserID   uint `json:"user_id" yaml:"user_id"`
+	ID       uint64 `json:"id" yaml:"id"`
+	Revision uint64 `json:"revision" yaml:"revision"`
+	UserID   uint64 `json:"user_id" yaml:"user_id"`
 }
